对用户锁定、删除后的逻辑做了处理。

+++ front/src/components/AppHeader.vue
@@ -2,7 +2,7 @@
  * @Author: J.sky bosichong@qq.com
  * @Date: 2022-11-29 19:48:29
  * @LastEditors: J.sky bosichong@qq.com
- * @LastEditTime: 2022-12-07 08:19:59
+ * @LastEditTime: 2022-12-13 00:02:46
  * @FilePath: /MiniAdmin/front/src/components/AppHeader.vue
 -->
 <template lang="">
@@ -11,21 +11,39 @@
     @back="() => null"
   >
     <template #extra >
-        <a-descriptions-item label="Created">欢迎您! {{username}} </a-descriptions-item>
-        <a-button @click="sessionStorageClear">退出</a-button>
+    
+      <a-dropdown>
+        <template #overlay>
+          <a-menu @click="handleMenuClick">
+            <a-menu-item key="1">修改资料</a-menu-item>
+            <a-menu-item key="2">修改密码</a-menu-item>
+            <a-menu-item key="3">退出</a-menu-item>
+          </a-menu>
+        </template>
+        <a-button >
+          欢迎您! {{username}}
+          <DownOutlined />
+        </a-button>
+      </a-dropdown>
+
+
     </template>
   
   </a-page-header>
 </template>
 <script setup>
 import { useRouter } from "vue-router";
-
+import { DownOutlined } from '@ant-design/icons-vue';
 const router = useRouter();
 const username = sessionStorage.getItem('username')
-const sessionStorageClear = (() => {
+
+const handleMenuClick = e => {
+  
+  if (e.key === '3') {
     sessionStorage.clear()
     router.push('/login')
-})
+  }
+};
 
 </script>
 <style lang="">
